Add necessary flags
I changed "g++"  to "gcc-8" as it  is the default brew alias  to the gnu
compiler. I  also added some lib  flags, by default mac  doesn't include
the "ncourses" so  I need to specify it, as  well as "lstdc++" (default)
and "-lc++"

"pty.h" Is not available on Mac so we should use "util.h" instead.

Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>